test: refactor user endpoint tests
