+++ bc_obps/registration/tests/endpoints/test_user_app_role_endpoints.py
@@ -1,0 +1,6 @@
+from registration.tests.utils.helpers import CommonTestSetup
+
+
+class TestUserAppRoleEndpoint(CommonTestSetup):
+    def test_get_user_app_role(self):
+        pass
